+++ src/routes/UserRecordContainer.js
@@ -40,7 +40,7 @@ class UserRecordContainer extends React.Component {
       path: 'groups',
       params: {
         query: 'cql.allRecords=1 sortby group',
-        limit: '40',
+        limit: '200',
       },
       records: 'usergroups',
     },