+++ src/components/UserDetailSections/PatronBlock/PatronBlock.js
@@ -4,7 +4,6 @@ import PropTypes from 'prop-types';
 import {
   FormattedMessage,
   injectIntl,
-  intlShape,
 } from 'react-intl';
 import moment from 'moment';
 
@@ -43,7 +42,7 @@ class PatronBlock extends React.Component {
     }),
     history: PropTypes.object,
     match: PropTypes.object,
-    intl: intlShape.isRequired,
+    intl: PropTypes.object.isRequired,
     onToggle: PropTypes.func,
     expanded: PropTypes.bool,
     accordionId: PropTypes.string,