+++ src/settings/PatronGroupsSettings.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import {
   FormattedMessage,
   injectIntl,
-  intlShape,
 } from 'react-intl';
 import { ControlledVocab } from '@folio/stripes/smart-components';
 import { stripesConnect, withStripes } from '@folio/stripes/core';
@@ -32,7 +31,7 @@ class PatronGroupsSettings extends React.Component {
     resources: PropTypes.shape({
       usersPerGroup: PropTypes.object,
     }).isRequired,
-    intl: intlShape.isRequired,
+    intl: PropTypes.object.isRequired,
   };
 
   constructor(props) {