+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-users
 
+## 3.0.3 (IN PROGRESS)
+
+* Increase limit of patron fee/fines, owners and patron groups. Refs UIU-1585.
+
 ## [3.0.2](https://github.com/folio-org/ui-users/tree/v3.0.2) (2020-04-09)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v3.0.1...v3.0.2)